CATALOG-3148: Move database trait to databases
