+++ query/src/catalogs/impls/database_catalog.rs
@@ -33,10 +33,10 @@ use common_meta_types::UpsertTableOptionReq;
 use crate::catalogs::catalog::Catalog;
 use crate::catalogs::impls::ImmutableCatalog;
 use crate::catalogs::impls::MutableCatalog;
-use crate::catalogs::Database;
 use crate::catalogs::Table;
 use crate::catalogs::TableFunction;
 use crate::configs::Config;
+use crate::databases::Database;
 use crate::table_functions::TableArgs;
 use crate::table_functions::TableFunctionFactory;
