+++ query/src/catalogs/impls/mutable_catalog.rs
@@ -39,11 +39,11 @@ use common_tracing::tracing;
 
 use crate::catalogs::backends::MetaRemote;
 use crate::catalogs::catalog::Catalog;
-use crate::catalogs::database::Database;
 use crate::catalogs::CatalogContext;
 use crate::catalogs::Table;
 use crate::common::MetaClientProvider;
 use crate::configs::Config;
+use crate::databases::Database;
 use crate::databases::DatabaseContext;
 use crate::databases::DatabaseFactory;
 use crate::storages::StorageContext;
