+++ query/src/storages/system/database.rs
@@ -15,9 +15,9 @@
 
 use std::sync::Arc;
 
-use crate::catalogs::Database;
 use crate::catalogs::InMemoryMetas;
 use crate::catalogs::Table;
+use crate::databases::Database;
 use crate::storages::system;
 
 #[derive(Clone)]
